Started work to unzip epub
Minizip is a library I am trying to use for unzipping the epub and getting the content file, which will let us identify the title, files containing pages, and the reading order of those pages. Once we can open those pages we can access the actual text and eventually parse it.